make Sockets a final class
