+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - `Innmind\OperatingSystem\Sockets::takeOver()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\OperatingSystem\Sockets::connectTo()` now returns an `Innmind\Immutable\Attempt`
 - `Innmind\OperatingSystem\OperatingSystem` is now a final class
+- `Innmind\OperatingSystem\Sockets` is now a final class
 
 ### Fixed
